Files for backend test project
Publishing all of the required files to create and modify dynamic entries on an HTML webpage.

generate_db.php is required to first run as a script or in browser to create all the MySQL data entries.

index.html relies on add.php, clear.php, and fetch.php to dynamically alter the webpage using AJAX.

Composer is used to load dependencies for PHPSpreadsheet for parsing through Excel files.